auth_service/requirements.txt
fastapi[standard]==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.1

python-keycloak==7.1.1
redis==5.0.1

PyYAML==6.0.2
